refactor: 중복되는 로직 제거 및 커스텀 훅 분리 #171
