+++ frontend/src/components/common/SettingModal/SettingModal.tsx
@@ -1,5 +1,4 @@
-import { useState } from 'react';
-
+import { useDropdown, useTimerPerRound, useTotalRound } from './SettingModal.hook';
 import {
   settingModalTitle,
   settingModalLayout,
@@ -15,42 +14,18 @@ import Modal from '../Modal/Modal';
 import { Category } from '@/types/room';
 
 const CATEGORY = ['음식', '연애', 'MBTI', '만약에'] as Category[];
+const TOTAL_ROUND_LIST = [5, 7, 10];
+const TIMER_PER_ROUND_LIST = [5, 10, 15];
 
 interface SettingModalProps {
   isOpen: boolean;
   onClose: () => void;
 }
 
-const useDropdown = () => {
-  const [category, setCategory] = useState('연애');
-
-  const handleClickOption = (e: React.MouseEvent<HTMLButtonElement>) => {
-    const target = e.target as HTMLButtonElement;
-    const clickedCategory = target.value;
-
-    if (!clickedCategory) return;
-    setCategory(clickedCategory);
-  };
-
-  return { category, handleClickOption };
-};
-
 const SettingModal = ({ isOpen, onClose }: SettingModalProps) => {
   const { category, handleClickOption } = useDropdown();
-  const [totalRound, setTotalRound] = useState(5);
-  const [timerPerRound, setTimerPerRound] = useState(10);
-
-  const handleClickRound = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    const target = e.target as HTMLButtonElement;
-
-    setTotalRound(Number(target.textContent));
-  };
-
-  const handleClickTimer = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    const target = e.target as HTMLButtonElement;
-
-    setTimerPerRound(Number(target.value));
-  };
+  const { totalRound, handleClickRound } = useTotalRound();
+  const { timerPerRound, handleClickTimer } = useTimerPerRound();
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} css={settingModalLayout}>
@@ -68,29 +43,30 @@ const SettingModal = ({ isOpen, onClose }: SettingModalProps) => {
             <span css={settingTitle}>총 라운드</span>
           </div>
           <div css={settingButtonContainer}>
-            <button css={settingButton(totalRound === 5)} onClick={handleClickRound}>
-              5
-            </button>
-            <button css={settingButton(totalRound === 7)} onClick={handleClickRound}>
-              7
-            </button>
-            <button css={settingButton(totalRound === 10)} onClick={handleClickRound}>
-              10
-            </button>
+            {TOTAL_ROUND_LIST.map((round) => (
+              <button
+                key={round}
+                css={settingButton(totalRound === round)}
+                onClick={handleClickRound}
+              >
+                {round}
+              </button>
+            ))}
           </div>
           <div css={settingTitleWrapper}>
             <span css={settingTitle}>라운드 당 타이머</span>
           </div>
           <div css={settingButtonContainer}>
-            <button css={settingButton(timerPerRound === 5)} onClick={handleClickTimer} value={5}>
-              5초
-            </button>
-            <button css={settingButton(timerPerRound === 10)} onClick={handleClickTimer} value={10}>
-              10초
-            </button>
-            <button css={settingButton(timerPerRound === 15)} onClick={handleClickTimer} value={15}>
-              15초
-            </button>
+            {TIMER_PER_ROUND_LIST.map((timer) => (
+              <button
+                key={timer}
+                css={settingButton(timerPerRound === timer)}
+                onClick={handleClickTimer}
+                value={timer}
+              >
+                {timer}초
+              </button>
+            ))}
           </div>
         </div>
       </Modal.Content>
